Started implementing tests.
