+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod rbackend;
 pub mod revent;
+pub mod rmercury;
 pub mod rplayer;
 pub mod rsystem;
 #[cfg(test)]
